pyproject.toml
[project]
name = "construction-leasing-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain>=0.3.26",
    "langchain-mcp-adapters>=0.1.7",
    "langchain-together>=0.3.0",
    "mcp[cli]>=1.10.1",
    "pandas>=2.3.0",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
    "streamlit>=1.46.0",
]
